make implicit typecast explicit
This came out of adding these compile options:
`-Werror -Wall -Wextra -Wshadow -Wconversion -Wno-deprecated-declarations -Wno-unused-parameter`

/src/monocypher.c
-u32  index     = lane * lane_size + (u32)ref;
+u32  index     = (u32)(lane * lane_size) + (u32)ref;

+++ src/monocypher.c
@@ -878,7 +878,7 @@ void crypto_argon2(u8 *hash, u32 hash_size, void *work_area,
 					u64  y         = (window_size * x) >> 32;
 					u64  z         = (window_size - 1) - y;
 					u64  ref       = (window_start + z) % lane_size;
-					u32  index     = lane * lane_size + (u32)ref;
+					u32  index     = (u32)(lane * lane_size) + (u32)ref;
 					blk *reference = blocks + index;
 
 					// Shuffle the previous & reference block
